% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackendData_ChromatinStates.R
\docType{data}
\name{BackendData_ChromatinStates}
\alias{BackendData_ChromatinStates}
\title{{Single Repositories of genomic features}}
\format{
An object of class \code{data.frame} with 2531 rows and 9 columns.
}
\usage{
BackendData_ChromatinStates
}
\description{
{
Integrative Fifteen-core chromatin state annotations across cell types and tissues.
}
}
\references{
{
1. Ernst, J. & Kellis, M. ChromHMM: automating chromatin-state discovery and characterization. Nat Methods 9, 215-6 (2012).
}
}
\keyword{datasets}
