% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackendData_GenomicFeatures.R
\docType{data}
\name{BackendData_GenomicFeatures}
\alias{BackendData_GenomicFeatures}
\title{{Single Repositories of genomic features}}
\format{
An object of class \code{data.frame} with 3057 rows and 7 columns.
}
\usage{
BackendData_GenomicFeatures
}
\description{
{
biomaRt repository cover 409,304 regulatory features including genomic coordination and associated sequence-motifs of 139729 CTCF binding site, 74575 enhancers, 63152 open chromatin region, 25313 promoters, 87975 promoter flanking region, and 18560 TF binding. UCSC repository provide annotations of 1,619,806 genomic features including 64334 3’UTRs, 114271 5’UTRs, 659198 introns, 289809 exons, and 82890 promoters. We downloaded the annotation databases from the biomaRt central porta (v0.6) (https://www.ensembl.org/biomart/) and TxDb.Hsapiens.UCSC.hg19.knownGene Annotation package.
}
}
\references{
{
1. Smedley, D. et al. The BioMart community portal: an innovative alternative to large, centralized data repositories. Nucleic Acids Res 43, W589-98 (2015).
}
}
\keyword{datasets}
