% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTL.R
\docType{data}
\name{eQTL}
\alias{eQTL}
\title{{Single eQTL profile}}
\format{
An object of class \code{data.frame} with 11715 rows and 21 columns.
}
\usage{
eQTL
}
\description{
{
eQTL profile including the following headers: seqnames, SNP_POS, SNP_POS, SNP_ID, gene_id.
}
}
\references{
{
1. Isar Nassiri, James Gilchrist, Evelyn Lau, Sara Danielli, Hussein Al Mossawi, JaneCheeseman, Matthew Neville, Julian C Knight, Benjamin P Fairfax.Genetic deter-minants of monocyte splicing are enriched for disease susceptibility loci includingfor COVID-19.
}
}
\keyword{datasets}
