% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEVV.R
\name{eSNPsEnrichmentAnalysis}
\alias{eSNPsEnrichmentAnalysis}
\title{Functional Enrichment of eQTL SNPs}
\usage{
eSNPsEnrichmentAnalysis(
  eQTL,
  TranscriptName,
  windowSize,
  FDRthreshold,
  BackendData_GenomicFeatures,
  BackendData_ChromatinStates,
  SNPs
)
}
\arguments{
\item{eQTL}{eQTL profile including the following headers: seqnames, SNP_POS, SNP_POS, SNP_ID, gene_id}

\item{TranscriptName}{Gene or isoform Ensembl ID}

\item{windowSize}{window around the TSS (e.g. 1000000)}

\item{FDRthreshold}{FDR threshold (e.g. 0.001)}

\item{BackendData_GenomicFeatures}{fifteen-core chromatin states from https://egg2.wustl.edu/roadmap/data/byFileType/chromhmmSegmentations/ChmmModels/coreMarks/jointModel/final/download/}

\item{BackendData_ChromatinStates}{10 genomic features from the biomaRt (Ensembl) and UCSC}

\item{SNPs}{List of all SNPs in human genome}
}
\value{
You can find the results in R object under title of 'RESULTsGenomicFeatures' and 'RESULTsChromatinState'.
}
\description{
eQTL SNPs as input: for eSNP enrichment analysis, first, for each gene/transcript, we selected associated eSNPs as foreground (F) and SNPs within the 1Mb window around the TSS as background (B) SNP sets. We used the number of overlaps of foreground (f) and background (b) SNP sets in the genomic feature or chromatin state and calculate the enrichment score (z-score and odds ratio).
}
\examples{
data(eQTL)
data(BackendData_GenomicFeatures)
data(BackendData_ChromatinStates)
data(SNPs)
eSNPsEnrichmentAnalysis(eQTL, TranscriptName = 'ENSG00000168310', windowSize=1000000, FDRthreshold = 0.001, BackendData_GenomicFeatures, BackendData_ChromatinStates, SNPs)
}
\author{
{Isar Nassiri, Benjamin Fairfax}
}
