% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPs.R
\docType{data}
\name{SNPs}
\alias{SNPs}
\title{{Single List of all SNPs in human genome}}
\format{
An object of class \code{data.frame} with 7217 rows and 11 columns.
}
\usage{
SNPs
}
\description{
{
List of all SNPs in human genome.
}
}
\references{
{
1. Isar Nassiri, James Gilchrist, Evelyn Lau, Sara Danielli, Hussein Al Mossawi, JaneCheeseman, Matthew Neville, Julian C Knight, Benjamin P Fairfax.Genetic deter-minants of monocyte splicing are enriched for disease susceptibility loci includingfor COVID-19.
}
}
\keyword{datasets}
